Refactor ContextLoaderListener from hotrod -> core
