+++ hotrod/pom.xml
@@ -53,10 +53,6 @@
             <artifactId>infinispan-client-hotrod</artifactId>
         </dependency>
         <dependency>
-            <groupId>org.springframework</groupId>
-            <artifactId>spring-web</artifactId>
-        </dependency>
-        <dependency>
             <groupId>org.wildfly</groupId>
             <artifactId>wildfly-clustering-infinispan-marshalling</artifactId>
         </dependency>
